Move YamlConfigProvider under com.uber.profiling
